Close the write/read disagreements the social goldens found
tests/packet_party_test.cpp, tests/packet_guild_test.cpp and
tests/packet_trade_test.cpp stated thirteen write/read disagreements as
tests that FAIL once the underlying code is fixed, and their review
reported two readers that could not be round-tripped at all. All of them
are fixed here and the flip-tests are rewritten as positive pins of the
behaviour the packets now produce, with new pins for the ones that only
became testable once fixed.

These packets are client-facing, so write() and getPacketSize() are the
contract: every fix is a refusal, a cap at a width the factory max
already budgets, a size-accounting correction, or a read-side correction
to what write() emits. No valid packet's bytes move: every golden .hex
file is unchanged. The two wire-layout movements are the guild roster
and union offer maxima below, and both are server-side read-buffer
budgets, not fields on the wire.

GuildInfo::write() emits a length byte and then the expiry date;
getSize() counted neither, so GCActiveGuildList declared a body a byte
plus the date shorter than the one it sends, for every guild in the
table, and writePacket() put that short length on the wire ahead of the
longer body. getSize() counts both now. getMaxSize() already budgeted
szBYTE + 11 for the date, so no maximum moves.

GCActiveGuildList::read(), GCGuildMemberList::read(),
GCWaitGuildList::read() and GCShowWaitGuildInfo::read() pushed each
record they parsed to the front of a list their write() emits front to
back, so the two guild tables, the roster and the founding-member list
all arrived reversed. All four push to the back now. Order is what the
player sees in each of them: the guild tables and the roster are lists
the client draws to pick from. Read side only; the wire does not move.

GCActiveGuildList::getListNum() and GCWaitGuildList::getListNum()
returned a BYTE of a list whose count goes on the wire as a WORD and
whose factory max budgets 5000 guilds, so a caller asking how many
guilds the packet holds was told zero at 256. Both return the WORD.

GCGuildMemberList, GCUnionOfferList, GCPartyJoined and GCTradeAddItem's
option and sub-item lists emitted every element they held against a
factory max that budgets a fixed number, so the body outgrew the read
buffer the receiver sizes from that max. Each refuses the entry past the
budget in addListElement, in write() and in read(), and destroys a
refused owned record, the way LCWorldList does with WorldInfo::kMaxCount.
Two of the maxima were wrong as well: GuildMemberInfo::getMaxSize() was
a whole 220-member table that counted one ServerID for all of them and
left out the roster's own count byte, and GCUnionOfferList's max was
exactly twenty offers with no room for the count byte in front of them.
The record maximum is one record now and each packet multiplies by its
own kMaxCount, which moves two lines in tests/wire-layout.txt:
GCGuildMemberList 5064 -> 5502 and GCUnionOfferList 1180 -> 1181. Of the
fill sites, two can reach a refusal: Guild::makeMemberInfo() walks the
guild's member map and nothing limits a guild to 220 members, and
GuildUnionOfferManager::makeOfferList() reads its offers from the
database. The party roster cannot -- PARTY_MAX_SIZE is the same six --
and the belt and armsband pocket counts in the item data stop at the
eight sub-items budgeted.

GCTradeAddItem kept its sub-item count in a member setListNum() wrote
and addListElement() left alone. write() put that member on the wire and
then emitted every record the list held, and getPacketSize() sized the
records from the member, so a packet built without a matching
setListNum() declared one length, announced a second count and emitted a
third number of records. The count is the list now and setListNum is
gone, the shape InventoryInfo, GearInfo, ExtraInfo and RideMotorcycleInfo
already have. Both callers -- the belt and the armsband branches of
CGTradeAddItemHandler::makeGCTradeAddItemPacket() -- set exactly the
number of records they had just added, so no count on the wire changes.

The protocol's strings were bounded on neither side. The guild and
member introductions in CGJoinGuild, CGModifyGuildIntro,
CGModifyGuildMemberIntro, CGRegistGuild, GCShowGuildInfo,
GCShowGuildMemberInfo and GCShowWaitGuildInfo each derived a BYTE length
from the string and emitted the string whole; the guards that looked
like a cap compared a BYTE against 255, which no BYTE can exceed. They
are cut to GUILD_INTRO_MAX_LENGTH in the setter and refused past it in
write(), the same width and shape as the inter-server introductions,
since every one of these budgets 255 or 256. The party names and chat
messages did the same against budgets of 10, 20 and 128: the names are
refused past their width, and when empty where the packet needs one,
while the messages are cut to 128 in the setter and refused when empty,
the way GCGuildChat already refuses its own. GCGuildChat's sending guild
name is refused empty or past GUILD_NAME_MAX_LENGTH, next to the sender
and the message it already bounded. A valid value is untouched, so the
wire does not move.

GCGuildResponse::getCode() and GCNPCResponse::getCode() returned a BYTE
of a member and a wire field that are both WORD, so a code past 255 --
GCNPCResponse declares 139 dialogue codes -- reached the wire whole and
the accessor whole halved. Both return the WORD; no server source calls
either accessor.

GCModifyGuildMemberInfo::read() and GCOtherGuildName::read() declared a
BYTE for the guild-name length, tested it against 30 and against 0, and
only then read the length the stream carries. So the read consumed a
length that was never checked when the indeterminate value happened to
be non-zero, and read no length byte at all when it happened to be zero,
leaving that byte in the stream and taking the rank from it. Both read
the length first and check it after, and both packets round trip now
instead of being pinned write-side only.

GCCreatureDied, GCDropItemToZone, GCMove, GCNPCSay and GCSkillFailed2
each carry a commented-out oStream.write((BYTE)48) and a read() that
consumes the byte it would have produced, the shape GCAddEffect and
GCAddMonsterFromBurrowing had: the reader took the first byte of the
object id as a flag and shifted every field after it. All five start at
the first field write() emits and the commented-out writes are gone.
GCDropItemToZone round trips through tests/packet_encrypter_test.cpp at
every encrypt code now; its goldens are unchanged, and the open finding
task 1.2 recorded for it is closed.

GCAddMonster, GCAddMonsterCorpse, GCAddMonsterFromBurrowing,
GCAddMonsterFromTransformation and PCVampireInfo3 indexed Dir2String
with a direction that arrives off the wire, so a value outside the eight
read past the end of the array. All five go through dir2String(), which
prints the number instead.

GCWaitGuildList and GCShowWaitGuildInfo join packet_guild_test.cpp with
the fix: they carry the same guild record and had the same reversal, and
their goldens are the two new files in the set.

+++ src/Core/GCNPCResponse.h
@@ -304,7 +304,7 @@ public:
     string toString() const;
 
 public:
-    BYTE getCode(void) const {
+    WORD getCode(void) const {
         return m_Code;
     }
     void setCode(WORD code) {
